Adding script that amends config automatically, reformatting config file, extending path checking for groups and buckets
Adding a script that's amended from Louise's manage_json script for her DynaFed LDAP plugin. This allows for automated adding and removing of group and bucket config information to the auth JSON file used for determining authorizational access to buckets. The script is designed to be used when groups are added/removed, buckets are added/removed, changes are made to which groups can access which buckets and with what permissions, etc. and will be usable in future updates of DynaFed with additional web functionality to allow for uploads of files, importing of buckets, etc.

The JSON file that contains the authorization config has been reformatted but keeps much of the original functionality. It allows for collating of buckets to groups and granular control of authorization to allow for specific groups to have specific permissions on specific buckets. With this model, a bucket can be mapped to multiple groups, with each group having a different level of access if required. Permissions for groups on all buckets belonging to them can be propogated down as well; for example if the example group has list permissions assigned to buckets it owns, then all buckets belonging to it can have their contents listed provided you are a member of the example group.

oidc_auth.py has been amended to no longer check for IP addresses. IP addresses are not relevant for this authorization project so are being removed. Minor additional changes to the path matching function take into account buckets and groups now, compared to a simple endpoint path that was in the previous solution.