Extending remove_group to remove group config file too, having auth JSON be re-read on every operation
Extending the functionality behind remove_group so that group.conf file is also removed and not just the information about the buckets and their config in the auth JSON file. This ensures all keys are wiped from the system as well.

For every operation performed on a bucket or file, oidc_auth.py reloads the auth JSON file to get the most up to date information on group permissions. This prevents permissions that have been updated or removed from not being reflected on the user's side as long as their IAM session continues. This replaces what previously happened, where the auth JSON was read into memory and remained consistent while the session lasted, which meant permissions could change while the user was interacting with buckets.