Removal of buckets from group config files and deletion of those files if empty afterwards
Extending the removal of buckets so that any groups that are empty with no buckets assigned to them are also removed from DynaFed. This includes removing them from the auth JSON and removing the group.conf file if it has no remaining buckets in it. This improves security by ensuring updated permissions on buckets are reflected universally immediately and prevents S3 keys from being left in DynaFed.