Adding additional checks for groups and buckets, implementing bucket permission amending
Adding function for checking if a group exists in the auth file
Adding function for checking if a bucket exists in the auth file
Adding function for updating permissions on a bucket. This works by receiving information on the bucket similar to adding a new bucket. It would typically be used when the user wants to update the groups who have read access and groups which have read/write access. Functions by removing the group from the auth file first and then adding it back in with the new settings. The config file for the group is unaffected during this operation.