Adding support for group names with forward slashes, fixing incorrect CORS configuration
Group names with slashes were causing issues when configuration files were generated, as code and config information contained pathnames, which complained when forward slashes were in group names. Fixing this by sanitising group names in appropriate places so that forward slashes are replaced with dashes '-'

Fixing CORS bug where buckets were being given an AllowedOrigin of the Echo URL. The point of this was misunderstood and it is now understood that AllowedOrigins specify where requests for bucket data come from, in this case the DynaFed websites. These are being added instead.