Adding bucket importing and removing functionality, requiring access and secret key for bucket removal
Including the scripts that trigger the importing and removing of buckets into DynaFed, called by the user through the web UI. Both require the group name, bucket name, access key and secret key for the sake of security. A successful operation returns a 201 or a 204 respectively, with a 409 being sent back in the event of a failure. This will become a more useful error message later on.

Additional modifications to manage_oidc_json.py allow for an extension of the sanitisation of group names that contain slashes by replacing them with dashes. It now applies to bucket removal code too.