Earlier check for existing bucket, removal of unnecessary parameter in request to import and remove
Calling the does_bucket_exist function in both import and remove bucket functions allows for a result to be returned back to the user earlier in the event of a failed attempt.

Also removed the unnecessary 'argument' parameter that was previously being sent to import.py and remove.py. This was leftover from a previous plan to call the script using the command line but is not needed as the correct function is invoked regardless.