Fixing bug in leftover print statement causing an error 500 on web frontend
In a scenario where the user entered garbage data into the import and removing forms on the web UI, this script was responding with an error 500 instead of the expected error 409. This was due to a print statement stating the issue when CORS rules were attempted to be updated on a non-existent bucket. The print statement was messing up the response to the web UI, meaning the calling script import.py/remove.py was crashing. Fixing this by ensuring nothing is printed in any operations called by the web UI. A basic "Operation failed" message is being added in the event that the script is called from the command line. More thorough error reporting will be implemented later.