Adding config file syncing between DynaFed hosts
This allows for consistency of available buckets between multiple DynaFed hosts. Useful in this case for hosts behind a load balancer. Uses a bucket on DynaFed dedicated specifically for holding bucket config files. This bucket is called s3://iris-dynafed-config and belongs to Sam Glendenning's S3 account, with the credentials being stored in a file called config-bucket-credentials.json. This file contains sensitive information, keys to the bucket, and so is stored in keys.gridpp.rl.ac.uk to be retrieved by the developer and placed in /etc/ugr/conf.d/

The bucket is designd to hold the correct and most up to date bucket configuration. Every 10 mins, a cron job on the host calls the get method in sync.py to get the latest list of files (in case one of the hosts has had an update to their config). When a change is made on a host, such as a new file, changed file or deleted file, the latest version of files is retrieved first, then the file is changed, then all the files are uploaded. If a file is deleted, it is deleted remotely first and then locally. More error checking will be coded later to account for changes made to config that disappeared in the last 10 mins, e.g. removing a bucket that has already been removed.

The script also ensures the owner of all the downloaded files is apache, otherwise the files may be unable to be changed by the apache user later.

The oidc_auth.json file is also stored in this bucket as this needs to be kept up to date and consistent as well.