Added comments, more error handling and renamed file
Added docstrings for all functions and general helpful comments to describe what's going on.

Added more error handling to allow errors to be indentified more easily. This includes more expansive error codes and messages for the user to receive on the web UI in the even of an error.

Renamed manage_oidc_json to manage_config as it is more appropriate for the job it performs.