Added some admin functionality, amended existing code to allow for admin privileges and admin verification
Allows admin to blacklist and whitelist buckets to prevent them from being imported by users. The ins and outs of this are still being worked on, such as how to inform users their buckets have been blacklisted, how users are informed about any imported buckets that have been blacklisted, if blacklisted buckets should be removed straight away etc.

remove.py and manage_config.py have been amended to account for an admin doing a simple bucket removal from a group. remove.py receives two variables from the web UI to verify this is an admin operation - a boolean set to true if a cookie is present claiming a currently logged in admin session (could be spoofed) and a list of the user's groups which should show they are a member of dynafed/admins group on IAM (cannot be spoofed). This prevents an attack where the user could simply pretend to be an admin by sending the right variables