Renaming admin to blacklisting, allowing admins to view any bucket, adding additional cgi-bin scripts for fetching groups and blacklisted buckets
Currently, this allows for admins to access any buckets regardless of whether or not admin mode is switched on on the front end. Hope to change this later on to take this into account.

Adding small scripts in cgi-bin to fetch a list of existing groups in DynaFed (used for admins removing buckets and populating the list of groups to take buckets from) and fetch a list of blacklisted buckets (for anyone to view what buckets aren't allowed).