Fixing bug in return statement, fixing bug in incorrect permissions on blacklist.json file
Previously, an error 500 was being returned from blacklist amendments when such an error hadn't actually occurred. This has been changed to the correct error code in line with the updated dictionary of return statements.

Also fixing bug where downloading the blacklist from the Echo bucket can cause it to have incorrect permissions (644). We need it to have 646 permissions, so os.chown is used for both the blacklist and oidc_auth.json files to ensure their permissions are correct upon download.