Download oidc_auth.json and blacklist.json from Echo upstream if files do not exist
When trying to open these files, previously either an error or an empty response was being returned if the file did not exist. This commit tries to fetch the file from the Echo upstream first and, if that fails, creates the file itself. This ensures it has the latest copy at all times.