Adding name of bucket for synchronising into config-bucket-credentials.json
Previously, the name of the bucket to synchronise .conf file was hard coded into sync.py. This has now been removed and the bucket name should now appear in config-bucket-credentials.json. This makes it easier for the name of the bucket to be changed if necessary, similar to the keys of the bucket itself.