Fixing bucket removal bug for non-admin users
Previously, users trying to remove a bucket that had already been deleted from Echo were unable to do so due to the credential check on the bucket failing (the bucket could no longer be tied to any credentials in Echo). This fix allows for users to pass another check whereby the bucket can be removed if the provided credentials match the ones supplied when the bucket was first imported (the ones that appear in the entry for the bucket in its group config file).